requirements.txt
Flask==2.0.1
flask_sqlalchemy==2.5.1
mysql-connector-python==8.0.23
google-cloud-vision==2.5.1
firebase-admin==5.2.0